Cleanup

+++ SmartFarmAndroidApp/app/src/main/java/com/example/smartfarmandroidapp/domain/CO2.java
@@ -2,13 +2,13 @@ package com.example.smartfarmandroidapp.domain;
 
 public class CO2 {
 
-    private int CO2;
+    private double CO2;
 
-    public CO2(int CO2) {
+    public CO2(double CO2) {
         this.CO2 = CO2;
     }
 
-    public int getCO2() {
+    public double getCO2() {
         return CO2;
     }
 }
